Move embeddings to top-level
