+++ packages/client/src/store/CommandsApi.ts
@@ -1,7 +1,10 @@
 import { ApiTopic, ClientBase } from "@vertesia/api-fetch-client";
-import { EmbeddingsStatusResponse, GenericCommandResponse, ProjectConfigurationEmbeddings, SupportedEmbeddingTypes } from "@vertesia/common";
+import { EmbeddingsApi } from "./EmbeddingsApi";
 
-
+/**
+ * @deprecated use EmbeddingsApi instead
+ * @see EmbeddingsApi
+ */
 export class CommandsApi extends ApiTopic {
 
     constructor(parent: ClientBase) {
@@ -9,34 +12,4 @@ export class CommandsApi extends ApiTopic {
     }
 
     embeddings = new EmbeddingsApi(this);
-
-}
-
-export class EmbeddingsApi extends ApiTopic {
-
-    constructor(parent: ClientBase) {
-        super(parent, "/embeddings");
-    }
-
-    async status(type: SupportedEmbeddingTypes): Promise<EmbeddingsStatusResponse> {
-        return this.get(type + "/status");
-    }
-
-    async activate(type: SupportedEmbeddingTypes, config: Partial<ProjectConfigurationEmbeddings>): Promise<GenericCommandResponse> {
-
-        if (!config.environment) {
-            throw new Error("Invalid configuration: select environment");
-        }
-
-        return this.post(type + "/enable", { payload: config });
-    }
-
-    async disable(type: SupportedEmbeddingTypes): Promise<GenericCommandResponse> {
-        return this.post(type + "/disable");
-    }
-
-    async recalculate(type: SupportedEmbeddingTypes): Promise<GenericCommandResponse> {
-        return this.post(type + "/recalculate");
-    }
-
 }
